Switches to Arrow function for meydaWA creation
Anonymous functions in JS dont maintain lexical scoping of the `this`
object. ES6 Arrow functions do. The problem was that `this` referred to
the function and not to the object, which was screwing up mel bands
internally. Could fix this the hard way, but we use Babel and ES6
anyway.