Drop support for Node 4 (EoL'd)
Doing this because after upgrading our dependencies we can't build meyda
on node 4 anymore because one of our dependencies has dropped support for
that. Before we release this, we should bump minor version (but we should
do that anyway because of the other changes we'll have merged before next
release.