chore: add commitlint (#392)
The idea here is that we always know whether a commit has an impact on the version (i.e. is a
breaking change, adds a feature, etc) so that we can automatically release a new version.

build: lint commits so that semantic-release works

Also, make sure that only the public npm registry is used, I kept pulling in from Spotify's private
registry which polluted the package-lock.json.