chore(workspaces): use npm workspaces to reduce deps churn
The showcase site has a subset of the dependencies of meyda, but since we weren't using workspaces
(because they were introduced recently), we had two dependabot installations to update each set of
deps independently. After confirming that dependencies of the workspace root are accessible to all
packages in a monorepo, I decided to remove all deps from the package, and stop independently
updating those deps.