fix(package exports): revert prettier rewrite
The prettier rewrite in #823 decided to export an es6 module rather than our default. I didn't
expect prettier to change behaviour, and I didn't catch it. This means that meyda@5.1.4 and
meyda@5.1.5 are bad, and will need to be deprecated. This commit introduces a test that will catch
this issue in the future.